drop dims type restriction

+++ src/layers/stateless.jl
@@ -44,13 +44,13 @@ logitbinarycrossentropy(logŷ, y) = (1 - y)*logŷ - logσ(logŷ)
 
     Normalises x to mean 0 and standard deviation 1, across the dimensions given by dims. Defaults to normalising over columns.
 """
-function normalise(x::AbstractArray; dims::Int=1)
+function normalise(x::AbstractArray; dims=1)
   μ′ = mean(x, dims = dims)
   σ′ = std(x, dims = dims, mean = μ′, corrected=false)
   return (x .- μ′) ./ σ′
 end
 
-function normalise(x::AbstractArray, dims::Int=1)
+function normalise(x::AbstractArray, dims=1)
   Base.depwarn("`normalise(x::AbstractArray, dims)` is deprecated, use `normalise(a, dims=dims)` instead.", :normalise)
   normalise(x, dims = dims)
 end
